requirements.txt
streamlit
opencv-python
pyttsx3
transformers
torch
Pillow
requests
llama-index-llms-upstage
tokenizers
mediapipe

